Moving SuccessPage to General Components
