+++ client/src/Utils/Routes/FundingRoutes.js
@@ -3,7 +3,7 @@ import React from 'react'
 // Pages
 import FundingAmount from '../../Components/Flows/FundingFlow/FundingAmount/FundingAmount';
 import FundingMethods from '../../Components/Flows/FundingFlow/FundingMethod/FundingMethods';
-import SuccessPage from '../../Components/Pages/CommonUse/SuccessPage/SuccessPage';
+import SuccessPage from '../../Components/General/SuccessPage';
 
 import AuthenticationRestrictedRoute from '../../Components/HOCs/AuthenticationRestrictedRoute'
