+++ client/src/Utils/Routes/TippingRoutes.js
@@ -3,7 +3,7 @@ import React from 'react'
 // Pages
 import SelectAmount from '../../Components/Flows/TippingFlow/SelectAmount/SelectAmount';
 import EnterReceiverCode from '../../Components/Flows/TippingFlow/EnterReceiverCode/EnterReceiverCode';
-import SuccessPage from '../../Components/Pages/CommonUse/SuccessPage/SuccessPage';
+import SuccessPage from '../../Components/General/SuccessPage';
 
 import AuthenticationRestrictedRoute from '../../Components/HOCs/AuthenticationRestrictedRoute'
